requirements.txt
asgiref==3.6.0
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cryptography==41.0.1
cymem==2.0.7
DAWG-Python==0.7.2
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
django-js-asset==2.0.0
django-tinymce==3.6.1
djangorestframework==3.14.0
docopt==0.6.2
flake8==6.0.0
graphviz==0.20.1
idna==3.4
Jinja2==3.1.2
jsmin==3.0.1
langcodes==3.3.0
MarkupSafe==2.1.2
mccabe==0.7.0
murmurhash==1.0.9
numpy==1.24.3
packaging==23.1
pathy==0.10.1
pdf2image==1.16.3
pdfminer==20191125
pdfminer.six==20221105
Pillow==9.5.0
preshed==3.0.8
psycopg==3.1.9
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.7
pydot==1.4.2
pyflakes==3.0.1
pygraphviz==1.10
pymorphy3==1.2.0
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.0.9
pytesseract==0.3.10
pytz==2023.3
requests==2.30.0
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.5.0/ru_core_news_sm-3.5.0-py3-none-any.whl#sha256=bedf427ab4f522137822d52c50ebdf9edeef31fdb808ed4776ce1596aa789704
smart-open==6.3.0
sorl-thumbnail==12.9.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==2.4.6
thinc==8.1.10
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==2.0.2
wasabi==1.1.1
